ORG-1129: exclure settings node while getting program

+++ webapp/src/main/java/org/exoplatform/community/brandadvocacy/portlet/frontend/JuZFrontEndApplication.java
@@ -311,7 +311,9 @@ public class JuZFrontEndApplication {
           }
         }
       }
-      this.jcrService.removeMissionParticipant(currentMissionParticipantId);
+      if (this.jcrService.removeMissionParticipantInParticipant(currentProgramId,remoteUserName,currentMissionParticipantId)) {
+        this.jcrService.removeMissionParticipant(currentMissionParticipantId);
+      }
       return Response.ok("nok");
     }
     else
